Split out profile dataclasses
